chore(engine): extract state inspecation logic for catch events
